services/radiologues/requirements.txt
fastapi[standard]==0.115.12
pymongo==4.5.0
python-dotenv==1.0.1
bcrypt==4.0.1
PyJWT==2.6.0
beautifulsoup4==4.9.3
schedule==1.1.0
requests==2.26.0
pillow==10.2.0
pytesseract==0.3.8
redis
pika
psutil
fpdf
prometheus_client
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-instrumentation-pymongo==0.52b0
opentelemetry-instrumentation-redis==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-instrumentation-grpc==0.52b0
opentelemetry-instrumentation==0.52b0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-semantic-conventions==0.52b0